split the IFMLLoadingPlugin from the main @Mod
