+++ src/main/java/squeek/applecore/AppleCore.java
@@ -1,18 +1,7 @@
 package squeek.applecore;
 
-import java.io.File;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-import net.minecraft.launchwrapper.Launch;
-
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
-import com.gtnewhorizon.gtnhmixins.IEarlyMixinLoader;
 
 import cpw.mods.fml.client.event.ConfigChangedEvent;
 import cpw.mods.fml.common.FMLCommonHandler;
@@ -22,7 +11,6 @@ import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 import cpw.mods.fml.common.event.FMLServerStartingEvent;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
-import cpw.mods.fml.relauncher.IFMLLoadingPlugin;
 import cpw.mods.fml.relauncher.Side;
 import squeek.applecore.api_impl.AppleCoreAccessorMutatorImpl;
 import squeek.applecore.api_impl.AppleCoreDispatcherImpl;
@@ -31,11 +19,8 @@ import squeek.applecore.client.DebugInfoHandler;
 import squeek.applecore.client.HUDOverlayHandler;
 import squeek.applecore.client.TooltipOverlayHandler;
 import squeek.applecore.commands.Commands;
-import squeek.applecore.mixinplugin.Mixins;
 import squeek.applecore.network.SyncHandler;
 
-@IFMLLoadingPlugin.SortingIndex(1100)
-@IFMLLoadingPlugin.MCVersion("1.7.10")
 @Mod(
         modid = ModInfo.MODID,
         name = ModInfo.MODID,
@@ -43,7 +28,7 @@ import squeek.applecore.network.SyncHandler;
         acceptableRemoteVersions = "*",
         guiFactory = ModInfo.GUI_FACTORY_CLASS,
         acceptedMinecraftVersions = "[1.7.10]")
-public class AppleCore implements IFMLLoadingPlugin, IEarlyMixinLoader {
+public class AppleCore {
 
     public static Logger Log = LogManager.getLogger(ModInfo.MODID);
 
@@ -53,7 +38,6 @@ public class AppleCore implements IFMLLoadingPlugin, IEarlyMixinLoader {
         AppleCoreAccessorMutatorImpl.values();
         AppleCoreDispatcherImpl.values();
         AppleCoreRegistryImpl.values();
-
         FMLCommonHandler.instance().bus().register(this);
     }
 
@@ -75,57 +59,5 @@ public class AppleCore implements IFMLLoadingPlugin, IEarlyMixinLoader {
     @SubscribeEvent
     public void onConfigChanged(ConfigChangedEvent.OnConfigChangedEvent event) {
         if (event.modID.equals(ModInfo.MODID)) ModConfig.sync();
-    }
-
-    @Override
-    public String getMixinConfig() {
-        return "mixins.AppleCore.early.json";
-    }
-
-    @Override
-    public List<String> getMixins(Set<String> loadedCoreMods) {
-        ModConfig.init(new File(Launch.minecraftHome, "config/" + ModInfo.MODID + ".cfg"));
-        // Manual coremod identification
-        try {
-            Class.forName("codechicken.lib.asm.ModularASMTransformer");
-            loadedCoreMods.add("codechicken.lib");
-        } catch (ClassNotFoundException ignored) {}
-
-        final List<String> mixins = new ArrayList<>();
-        final List<String> notLoading = new ArrayList<>();
-        for (Mixins mixin : Mixins.values()) {
-            if (mixin.phase == Mixins.Phase.EARLY) {
-                if (mixin.shouldLoad(loadedCoreMods, Collections.emptySet())) {
-                    mixins.addAll(mixin.mixinClasses);
-                } else {
-                    notLoading.addAll(mixin.mixinClasses);
-                }
-            }
-        }
-        Log.info("Not loading the following EARLY mixins: {}", notLoading.toString());
-        return mixins;
-    }
-
-    @Override
-    public String[] getASMTransformerClass() {
-        return new String[0];
-    }
-
-    @Override
-    public String getModContainerClass() {
-        return null;
-    }
-
-    @Override
-    public String getSetupClass() {
-        return null;
-    }
-
-    @Override
-    public void injectData(Map<String, Object> data) {}
-
-    @Override
-    public String getAccessTransformerClass() {
-        return null;
     }
 }
