+++ src/main/java/squeek/applecore/mixinplugin/AppleCoreLoadingPugin.java
@@ -1,0 +1,71 @@
+package squeek.applecore.mixinplugin;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import net.minecraft.launchwrapper.Launch;
+
+import com.gtnewhorizon.gtnhmixins.IEarlyMixinLoader;
+
+import cpw.mods.fml.relauncher.IFMLLoadingPlugin;
+import squeek.applecore.ModConfig;
+import squeek.applecore.ModInfo;
+
+@IFMLLoadingPlugin.MCVersion("1.7.10")
+public class AppleCoreLoadingPugin implements IFMLLoadingPlugin, IEarlyMixinLoader {
+
+    @Override
+    public String[] getASMTransformerClass() {
+        return null;
+    }
+
+    @Override
+    public String getModContainerClass() {
+        return null;
+    }
+
+    @Override
+    public String getSetupClass() {
+        return null;
+    }
+
+    @Override
+    public void injectData(Map<String, Object> data) {}
+
+    @Override
+    public String getAccessTransformerClass() {
+        return null;
+    }
+
+    @Override
+    public String getMixinConfig() {
+        return "mixins.AppleCore.early.json";
+    }
+
+    @Override
+    public List<String> getMixins(Set<String> loadedCoreMods) {
+        ModConfig.init(new File(Launch.minecraftHome, "config/" + ModInfo.MODID + ".cfg"));
+        // Manual coremod identification
+        try {
+            Class.forName("codechicken.lib.asm.ModularASMTransformer");
+            loadedCoreMods.add("codechicken.lib");
+        } catch (ClassNotFoundException ignored) {}
+
+        final List<String> mixins = new ArrayList<>();
+        final List<String> notLoading = new ArrayList<>();
+        for (Mixins mixin : Mixins.values()) {
+            if (mixin.phase == Mixins.Phase.EARLY) {
+                if (mixin.shouldLoad(loadedCoreMods, Collections.emptySet())) {
+                    mixins.addAll(mixin.mixinClasses);
+                } else {
+                    notLoading.addAll(mixin.mixinClasses);
+                }
+            }
+        }
+        return mixins;
+    }
+}
